% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdocsyntax.r
\name{enable_html_doc_highlight}
\alias{enable_html_doc_highlight}
\title{Enable syntax highlighting in R html doc}
\usage{
enable_html_doc_highlight(set_html_help_type = FALSE)
}
\arguments{
\item{set_html_help_type}{Automatically set the \code{help_type} option to \code{html},
i.e. \code{options(help_type = "html")}. Default is \code{FALSE}.}
}
\value{
The original browser
}
\description{
Enable syntax highlighting in R html doc
}
\details{
R html docs are displayed by running the doc url through a browser function.
The browser in use is stored in the \code{browser} option, \code{getOption("browser")}.
This package works by creating a browser function that takes in the R doc html,
applies syntax highlighting, then displays it using the previously set browser, or,
in case RStudio is running, \code{\link[rstudioapi]{viewer}}.
}
\examples{

\dontrun{
# minimal example, using RStudio
# original R html doc without syntax highlighting
?sapply
# enable syntax highlighting
enable_html_doc_highlight()
?sapply
}

\dontrun{
# this works if not using RStudio
# set html doc, alternatively can use help(..., help_type = "html")
options(help_type = "html")
# original R html doc without syntax highlighting
?sapply
# enable syntax highlighting
enable_html_doc_highlight()
?sapply
}

\dontrun{
# in case you want to be able switch off syntax highlighting,
# i.e. revert to original state

# store the original browser
original_browser <- enable_html_doc_highlight()
# revert to the original browser
options(browser = original_browser)
}

}
\seealso{
\code{\link[utils]{browseURL}}
}
