% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdocsyntax.r
\name{highlight_html_docs}
\alias{highlight_html_docs}
\title{Enable syntax highlighting in R html doc}
\usage{
highlight_html_docs()
}
\description{
Enable syntax highlighting in R html doc
}
\details{
R html docs are displayed by running the doc url through a browser function.
The browser in use is stored in the \code{browser} option, \code{getOption("browser")}.
This package works by creating a browser function that takes in the R doc html,
applies syntax highlighting, then displays it using the previously set browser, or,
in case RStudio is running, \code{\link[rstudioapi]{viewer}}.
}
\examples{

\dontrun{
# minimal example, using RStudio
# original R html doc without syntax highlighting
?sapply

# enable syntax highlighting
highlight_html_docs()

# Code in HTML documents is now highlighted
?sapply
# or if help pages are not displayed in HTML mode by default,
# e.g. when R is not running inside RStudio
help(sapply, help_type = "html")

#' # Disable syntax highlighting
unhighlight_html_docs()
}

}
\seealso{
\code{\link[utils]{browseURL}}
}
