% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{read_text}
\alias{read_text}
\title{Read a file into a string character.
Similar to \code{readr::read_file}.}
\usage{
read_text(f)
}
\description{
Read a file into a string character.
Similar to \code{readr::read_file}.
}
