% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{tempf}
\alias{tempf}
\title{Create a temp file preloaded with content and return the file path.}
\usage{
tempf(content, ...)
}
\description{
Create a temp file preloaded with content and return the file path.
}
