% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdocsyntax.r
\name{use_highlight_browser}
\alias{use_highlight_browser}
\title{Enable syntax highlighting in R doc viewed with RStudio}
\usage{
use_highlight_browser()
}
\description{
Enable syntax highlighting in R doc viewed with RStudio
}
